move classes around to make docs easier to read
