+++ flare/components/functional.py
@@ -2,25 +2,23 @@ from __future__ import annotations
 
 import abc
 import typing as t
-import typing_extensions as te
-import hikari
-import sigparse
-from flare import dataclass
 
+import sigparse
+import typing_extensions as te
+
+from flare import dataclass
 from flare.components import base
-from flare.components.button import Button
-from flare.components.select import Select
 
 if t.TYPE_CHECKING:
     from flare import context
 
-__all__: t.Sequence[str] = ("button", "select")
+__all__: t.Sequence[str] = ("FunctionalComponent",)
 
 P = te.ParamSpec("P")
 T = t.TypeVar("T", bound="base.Component")
 
 
-class _FunctionalComponent(abc.ABC, t.Generic[T]):
+class FunctionalComponent(abc.ABC, t.Generic[T]):
     """
     Decorator to wrap a component function callback so it can be treated as a
     class internally. This should be inherited to create decorators for
@@ -66,81 +64,6 @@ class _FunctionalComponent(abc.ABC, t.Generic[T]):
     @abc.abstractmethod
     def kwargs(self) -> dict[str, t.Any]:
         """The kwargs for `__init_subclass__`"""
-
-
-class button(_FunctionalComponent[Button]):
-    """
-    A decorator to create a `flare.Button`. This is a shorthand for when type
-    safety is not needed.
-    """
-
-    def __init__(
-        self,
-        *,
-        cookie: str | None = None,
-        label: str | None = None,
-        emoji: str | hikari.Emoji | None = None,
-        style: hikari.ButtonStyle = hikari.ButtonStyle.PRIMARY,
-        disabled: bool = False,
-    ) -> None:
-        self.cookie = cookie
-        self.label = label
-        self.emoji = emoji
-        self.style = style
-        self.disabled = disabled
-
-    @property
-    def component_type(self) -> type[Button]:
-        return Button
-
-    @property
-    def kwargs(self) -> dict[str, t.Any]:
-        return {
-            "cookie": self.cookie,
-            "label": self.label,
-            "emoji": self.emoji,
-            "style": self.style,
-            "disabled": self.disabled,
-        }
-
-
-class select(_FunctionalComponent[Select]):
-    """
-    A decorator to create a `flare.Select`. This is a shorthand for when type
-    safety is not needed.
-    """
-
-    def __init__(
-        self,
-        *,
-        cookie: str | None = None,
-        options: t.Sequence[tuple[str, str] | str] | None = None,
-        min_values: int | None = None,
-        max_values: int | None = None,
-        placeholder: hikari.UndefinedOr[str] = hikari.UNDEFINED,
-        disabled: bool | None = None,
-    ) -> None:
-        self.cookie = cookie
-        self.options = options
-        self.min_values = min_values
-        self.max_values = max_values
-        self.placeholder = placeholder
-        self.disabled = disabled
-
-    @property
-    def component_type(self) -> type[Select]:
-        return Select
-
-    @property
-    def kwargs(self) -> dict[str, t.Any]:
-        return {
-            "cookie": self.cookie,
-            "options": self.options,
-            "min_values": self.min_values,
-            "max_values": self.max_values,
-            "placeholder": self.placeholder,
-            "disabled": self.disabled,
-        }
 
 
 # MIT License
